made a lot of changes for stream to work

+++ src/schema.py
@@ -10,5 +10,5 @@ INPUT_SCHEMA = {
     "max_new_tokens": {"type": int, "required": False, "default": 1024},
     "token_repetition_penalty": {"type": float, "required": False, "default": 1.15},
     "token_repetition_range": {"type": int, "required": False, "default": -1},
-    "token_repetition_decay": {"type": float, "required": False, "default": 0},
+    "token_repetition_decay": {"type": int, "required": False, "default": 0},
 }
